+++ mod-inventory-storage-server/src/main/java/org/folio/rest/impl/InitApiImpl.java
@@ -15,20 +15,44 @@ import org.folio.services.consortium.ServicePointSynchronizationVerticle;
 import org.folio.services.consortium.ShadowInstanceSynchronizationVerticle;
 import org.folio.services.consortium.SynchronizationVerticle;
 import org.folio.services.migration.async.AsyncMigrationConsumerVerticle;
+import org.folio.services.s3storage.FolioS3ClientFactory;
+import org.folio.services.s3storage.FolioS3ClientFactory.S3ConfigType;
+import org.folio.utils.Environment;
 
 public class InitApiImpl implements InitAPI {
 
   private static final Logger log = LogManager.getLogger();
 
+  private static final String S3_CONFIG_REQUIRED_KEY = "S3_CONFIG_REQUIRED";
+
   @Override
   public void init(Vertx vertx, Context context, Handler<AsyncResult<Boolean>> handler) {
     initConsortiumDataCache(vertx, context);
-    initAsyncMigrationVerticle(vertx)
+    validateS3Configurations()
+      .compose(v -> initAsyncMigrationVerticle(vertx))
       .compose(v -> initShadowInstanceSynchronizationVerticle(vertx, getConsortiumDataCache(context)))
       .compose(v -> initSynchronizationVerticle(vertx, getConsortiumDataCache(context)))
       .compose(v -> initServicePointSynchronizationVerticle(vertx, getConsortiumDataCache(context)))
       .map(true)
       .onComplete(handler);
+  }
+
+  private Future<Void> validateS3Configurations() {
+    var isS3ConfigRequired = Environment.getBoolValue(S3_CONFIG_REQUIRED_KEY, Boolean.FALSE);
+    if (Boolean.FALSE.equals(isS3ConfigRequired)) {
+      return Future.succeededFuture();
+    }
+    try {
+      for (S3ConfigType configType : S3ConfigType.values()) {
+        var folioS3Client = FolioS3ClientFactory.getFolioS3Client(configType);
+        folioS3Client.createBucketIfNotExists();
+        log.info("validateS3Configurations:: S3 client configured successfully for config type {}", configType);
+      }
+      return Future.succeededFuture();
+    } catch (Exception e) {
+      log.error("init:: S3 configuration validation failed", e);
+      return Future.failedFuture(e);
+    }
   }
 
   private Future<Void> initAsyncMigrationVerticle(Vertx vertx) {